Rename FlowDiagram to FlowsView and remove FlowNode widget
