--- supervisely/app/widgets/flow_node/template.html
@@ -1,3 +1,0 @@
-<div>
-    widget.content
-</div>
